requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
bcc==0.18.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.1.8
colorama==0.4.4
Command==0.1.0
command-not-found==0.3
config==0.5.1
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==1.1.2
h11==0.16.0
hidpidaemon==18.4.6
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==8.10.0
msgpack==1.0.3
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
orjson==3.10.18
pop-transition==1.1.2
protobuf==3.12.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycairo==1.20.1
pycups==2.0.1
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydbus==0.6.0
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.0
python-gnupg==0.4.8
python-jose==3.4.0
python-multipart==0.0.20
python-slugify==8.0.4
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==6.1.0
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
rich==14.0.0
rich-toolkit==0.14.6
rsa==4.9.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
systemd-python==234
text-unidecode==1.3
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.10.0
urllib3==1.26.5
uvicorn==0.34.2
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.0.5
websockets==15.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
